requirements.txt
Flask
gunicorn
python-firebase